refactor: rename files to use hyphen
