+++ src/react/table-app/function-cell/general-function.ts
