+++ src/react/table-app/function-cell/number-function.ts
